Fix : getAPItoken 함수를 lib > api 폴더로 이동
