+++ src/app/api/generate/route.ts
@@ -1,6 +1,6 @@
+import { getAPItoken } from "@/lib/api/llama3";
 import { LLAMA_API_URL } from "@/lib/const";
 import { NextRequest, NextResponse } from "next/server";
-import { getAPItoken } from "../authenticate/route";
 
 export async function POST(request: NextRequest) {
   try {
